Added a version constant to constants.py.
Added some code to size() so that small canvases are surrounded by a 
margin so as to get around a problem with small windows in MS Windows 
(is this why Processing does the same thing? I wonder...)

Ellipses now use a smaller sampling in order to help speed up some 
scripts.